+++ src/main/java/de/jeffclan/JeffChestSort/JeffChestSortPlugin.java
@@ -33,7 +33,6 @@ import java.io.BufferedWriter;
 */
 
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.FilenameFilter;
 import java.io.IOException;
@@ -358,6 +357,7 @@ public class JeffChestSortPlugin extends JavaPlugin {
 
 	private void registerMetrics() {
 		// Metrics will need json-simple with 1.14 API.
+		//Metrics bStats = new Metrics(this,3089); // Metrics will need plugin ID with bStats 1.7+
 		Metrics bStats = new Metrics(this);
 		bStats.addCustomChart(new Metrics.SimplePie("sorting_method", () -> sortingMethod));
 		bStats.addCustomChart(new Metrics.SimplePie("config_version",